Try to do first programm 19

+++ lab_04_2_1/main.c
@@ -2,12 +2,12 @@
 
 void srar(int array[], int n, int rs2)
 {
-    double s = 0;
+    float s = 0;
     for (int i = 0; i < n; i++)
         if (array[i] < 0)
             s += array[i];
     s = s / rs2;
-    printf("%lf", s);
+    printf("%f", s);
 }
 
 int main(void)
